movie_project/backend/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
requests==2.26.0  # Совместимая версия с kinopoisk-api-unofficial-client